refactor: Use `sequence` in more places
Plus some general clearing up

Code-quality-wise `prqlc-parser` in a pretty good place now, I think...
