+++ Cargo.toml
@@ -56,3 +56,5 @@ semver = {version = "1.0.23", features = ["serde"]}
 serde = {version = "1.0.204", features = ["derive"]}
 serde_json = "1.0.120"
 serde_yaml = {version = "0.9.34"}
+similar = "2.5.0"
+similar-asserts = "1.5.0"
