+++ prqlc/prqlc-parser/Cargo.toml
@@ -35,3 +35,5 @@ chumsky = {version = "0.9.2", features = ["ahash", "std"], default-features = fa
 [dev-dependencies]
 insta = {workspace = true}
 serde_json = {workspace = true}
+similar = {workspace = true}
+similar-asserts = {workspace = true}
